major refactor: add cli options and use httpx instead of requests
This is basically a rewrite of everything except extracting links from HTML.
link_checker now uses httpx and it's `mounts` system to implement
transparently accessing the filesystem instead of a remote server. This
avoids the somewhat troublesome http server on localhost:8000 used before.
The concurrency implementation using threads is also gone in favor of
asyncio and httpx'es async support.

To more cleanly support additional features argparse is now used to
parse arguments from the commandline. One of those is an easy way
to ignore URLs.